+++ test/sql/other.sql
@@ -12,10 +12,10 @@ select dl_chattr('public','sample_datalinks2','link', dl_lco(link_control=>'FILE
 
 select sql_identifier,lco,dlco.*
   from datalink.dl_columns left join datalink.dl_link_control_options dlco using (lco);
-
+/*
 insert into sample_datalinks2 (link)
-values (dlvalue('http://www.ljudmila.org','URL','Sample datalink'));
-
+values (dlvalue('http://en.wikipedia.org','URL','Sample datalink'));
+*/
 insert into sample_datalinks2 (link)
 values (dlvalue('/etc/','FS','Sample file datalink'));